RavenDB-24372 Styling
