+++ src/Raven.Studio/typescript/components/common/aiAssistant/AiAssistantEulaModal.tsx
